LANG-1733: Move null handling features to Consumers and Functions
