+++ src/main/java/org/apache/commons/lang3/function/Consumers.java
@@ -21,7 +21,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 /**
- * Provides {@link Consumer} instances.
+ * Provides {@link Consumer} instances and utilities for working with {@link Consumer}
  *
  * @since 3.13.0
  */
@@ -46,4 +46,17 @@ public class Consumers {
         // No instances.
     }
 
+    /**
+     * Applies the given {@link Consumer} action to the object if the object is not null. If the object is null, it does
+     * nothing.
+     *
+     * @param object the object to be consumed
+     * @param consumer the consumer to consume
+     * @param <T> the type of the argument the consumer accepts
+     */
+    public static <T> void acceptIfNotNull(final T object, final Consumer<T> consumer) {
+        if (object != null) {
+            consumer.accept(object);
+        }
+    }
 }
