--- src/test/java/org/apache/commons/lang3/function/NullSafeOperationsTest.java
@@ -1,41 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.commons.lang3.function;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class NullSafeOperationsTest {
-
-    @Test
-    public void testApplyIfNotNull() {
-        assertEquals("jay-bae", NullSafeOperations.applyIfNotNull("jay", string -> string.concat("-bae")));
-        assertDoesNotThrow(() -> NullSafeOperations.applyIfNotNull((String) null, string -> fail()));
-        assertNull(NullSafeOperations.applyIfNotNull((String) null, string -> string.concat("abc")));
-    }
-
-    @Test
-    public void testAcceptIfNotNull() {
-        StringBuilder builder = new StringBuilder("jay");
-        NullSafeOperations.acceptIfNotNull(builder, sb -> sb.append("-bae"));
-        assertEquals("jay-bae", builder.toString());
-
-        assertDoesNotThrow(() -> NullSafeOperations.acceptIfNotNull((String) null, string -> fail()));
-    }
-}
